saving becomes better (there was a bug before now). Time to push to master.
Master Status: baseline works well. But that's all we've tried. We don't know if Pruner does anything. The NER eval is clearly not working. We also haven't put HOI pruner here just now. That's also something to do. There also seems to be some memory stufff going on. The memory usage is all over the place and to be frank its getting irritating now.

On another branch, we fork out to fix the resume code.